Split CLI into multiple binaries
By consolidating all CLI commands into a single binary, we gain the following benefits:

- `cargo check --workspace` now passes as previously one would be
  required to enable at least one feature from the cli crate
- Fewer combinations of features to test (like when linting).
