+++ Cargo.toml
@@ -5,12 +5,14 @@ members = [
     "src/applib",
     "src/bumpalo-serde-derive",
     "src/bumpalo-serde",
-    "src/cli",
     "src/eu4game-data",
     "src/eu4game",
+    "src/pdx-admin-cli",
     "src/pdx-assets",
+    "src/pdx-fun-cli",
     "src/pdx-map-cli",
     "src/pdx-map",
+    "src/pdx-tokenize-cli",
     "src/pdx-zstd",
     "src/rawbmp",
     "src/schemas",
