--- src/cli/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "pdx"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[lints]
-workspace = true
-
-[features]
-admin = ["tokenize", "dep:applib", "dep:eu4game", "dep:eu4save", "dep:flate2", "dep:pdx-zstd"]
-fun = ["dep:applib", "dep:eu4game", "dep:eu4save", "dep:flate2", "dep:rawbmp", "dep:pdx-assets", "dep:schemas"]
-
-tokenize = ["dep:pdx-zstd", "dep:schemas"]
-
-[dependencies]
-anyhow = { workspace = true }
-applib = { optional = true , workspace = true, default-features = false, features = ["zstd_c"] }
-bytemuck = { workspace = true }
-clap = { features = ["derive", "env"] , workspace = true }
-csv = { workspace = true }
-env_logger = { workspace = true }
-eu4game = { optional = true, default-features = false , workspace = true, features = ["zstd_c"] }
-eu4save = { default-features = false, features = ["serialize", "zstd_c"], optional = true , workspace = true }
-flate2 = { optional = true, default-features = false, features = ["zlib-rs"] , workspace = true }
-highway = { workspace = true }
-log = { workspace = true }
-pdx-assets = { workspace = true, optional = true }
-pdx-map = { workspace = true, default-features = false }
-pdx-zstd = { workspace = true, optional = true, default-features = false, features = ["zstd_c"] }
-rawbmp = { workspace = true, optional = true }
-rawzip = { workspace = true }
-rayon = { workspace = true }
-schemas = { optional = true , workspace = true }
-serde = { features = ["derive"] , workspace = true }
-serde_json = { workspace = true }
-walkdir = { workspace = true }
-
-[target.'cfg(target_env = "musl")'.dependencies]
-mimalloc = { workspace = true }
