--- src/cli/src/cli/mod.rs
@@ -1,3 +1,0 @@
-mod verbosity;
-
-pub use verbosity::*;
