--- src/cli/src/cli/verbosity.rs
@@ -1,140 +1,0 @@
-/* ref: https://github.com/rust-cli/clap-verbosity-flag */
-#![allow(dead_code)]
-
-#[derive(clap::Args, Debug, Clone)]
-pub struct Verbosity<L: LogLevel = ErrorLevel> {
-    #[arg(
-        long,
-        short = 'v',
-        action = clap::ArgAction::Count,
-        global = true,
-        help = L::verbose_help(),
-        long_help = L::verbose_long_help(),
-    )]
-    verbose: u8,
-
-    #[arg(
-        long,
-        short = 'q',
-        action = clap::ArgAction::Count,
-        global = true,
-        help = L::quiet_help(),
-        long_help = L::quiet_long_help(),
-        conflicts_with = "verbose",
-    )]
-    quiet: u8,
-
-    #[arg(skip)]
-    phantom: std::marker::PhantomData<L>,
-}
-
-impl<L: LogLevel> Verbosity<L> {
-    /// Create a new verbosity instance by explicitly setting the values
-    pub fn new(verbose: u8, quiet: u8) -> Self {
-        Verbosity {
-            verbose,
-            quiet,
-            phantom: std::marker::PhantomData,
-        }
-    }
-
-    /// Get the log level.
-    ///
-    /// `None` means all output is disabled.
-    pub fn log_level(&self) -> Option<log::Level> {
-        level_enum(self.verbosity())
-    }
-
-    /// Get the log level filter.
-    pub fn log_level_filter(&self) -> log::LevelFilter {
-        level_enum(self.verbosity())
-            .map(|l| l.to_level_filter())
-            .unwrap_or(log::LevelFilter::Off)
-    }
-
-    /// If the user requested complete silence (i.e. not just no-logging).
-    pub fn is_silent(&self) -> bool {
-        self.log_level().is_none()
-    }
-
-    fn verbosity(&self) -> u8 {
-        level_value(L::default()) - self.quiet + self.verbose
-    }
-}
-
-fn level_value(level: Option<log::Level>) -> u8 {
-    match level {
-        None => 0,
-        Some(log::Level::Error) => 1,
-        Some(log::Level::Warn) => 2,
-        Some(log::Level::Info) => 3,
-        Some(log::Level::Debug) => 4,
-        Some(log::Level::Trace) => 5,
-    }
-}
-
-fn level_enum(verbosity: u8) -> Option<log::Level> {
-    match verbosity {
-        0 => None,
-        1 => Some(log::Level::Error),
-        2 => Some(log::Level::Warn),
-        3 => Some(log::Level::Info),
-        4 => Some(log::Level::Debug),
-        _ => Some(log::Level::Trace),
-    }
-}
-
-use std::fmt;
-
-impl<L: LogLevel> fmt::Display for Verbosity<L> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.verbosity())
-    }
-}
-
-pub trait LogLevel {
-    fn default() -> Option<log::Level>;
-
-    fn verbose_help() -> Option<&'static str> {
-        Some("More output per occurrence")
-    }
-
-    fn verbose_long_help() -> Option<&'static str> {
-        None
-    }
-
-    fn quiet_help() -> Option<&'static str> {
-        Some("Less output per occurrence")
-    }
-
-    fn quiet_long_help() -> Option<&'static str> {
-        None
-    }
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct ErrorLevel;
-
-impl LogLevel for ErrorLevel {
-    fn default() -> Option<log::Level> {
-        Some(log::Level::Error)
-    }
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct WarnLevel;
-
-impl LogLevel for WarnLevel {
-    fn default() -> Option<log::Level> {
-        Some(log::Level::Warn)
-    }
-}
-
-#[derive(Copy, Clone, Debug, Default)]
-pub struct InfoLevel;
-
-impl LogLevel for InfoLevel {
-    fn default() -> Option<log::Level> {
-        Some(log::Level::Info)
-    }
-}
