+++ src/pdx-admin-cli/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "pdx-admin-cli"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[[bin]]
+name = "pdx-admin"
+path = "src/main.rs"
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+applib = { workspace = true, default-features = false, features = ["zstd_c"] }
+bytemuck = { workspace = true }
+clap = { features = ["derive", "env"], workspace = true }
+csv = { workspace = true }
+env_logger = { workspace = true }
+eu4game = { workspace = true, default-features = false, features = ["zstd_c"] }
+eu4save = { workspace = true, default-features = false, features = ["serialize", "zstd_c"] }
+flate2 = { workspace = true, default-features = false, features = ["zlib-rs"] }
+highway = { workspace = true }
+log = { workspace = true }
+pdx-map = { workspace = true, default-features = false }
+pdx-zstd = { workspace = true, default-features = false, features = ["zstd_c"] }
+rawzip = { workspace = true }
+rayon = { workspace = true }
+schemas = { workspace = true }
+serde = { features = ["derive"], workspace = true }
+serde_json = { workspace = true }
+walkdir = { workspace = true }
+
+[target.'cfg(target_env = "musl")'.dependencies]
+mimalloc = { workspace = true }
