+++ src/pdx-admin-cli/src/color_mapper.rs
