+++ src/pdx-fun-cli/src/terrain_mapper.rs
@@ -1,0 +1,20 @@
+use clap::Args;
+use std::{path::PathBuf, process::ExitCode};
+
+/// Terrain mapper (currently disabled - implementation commented out)
+#[derive(Args)]
+pub struct TerrainMapperArgs {
+    /// Map directory (currently unused)
+    #[clap(value_parser)]
+    _map_dir: PathBuf,
+}
+
+impl TerrainMapperArgs {
+    pub fn run(&self) -> anyhow::Result<ExitCode> {
+        eprintln!("TerrainMapper is currently disabled. The implementation is commented out.");
+        Ok(ExitCode::FAILURE)
+    }
+}
+
+// Original implementation is preserved below but commented out:
+// [... all the original commented code would go here ...]
